+++ modules/server/src/main/java/io/jpom/socket/handler/BaseTerminalHandler.java
@@ -22,8 +22,8 @@
  */
 package io.jpom.socket.handler;
 
-import cn.hutool.core.util.StrUtil;
 import io.jpom.socket.BaseHandler;
+import io.jpom.util.SocketSessionUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.socket.BinaryMessage;
 import org.springframework.web.socket.WebSocketSession;
@@ -37,22 +37,15 @@ import java.io.IOException;
 @Slf4j
 public abstract class BaseTerminalHandler extends BaseHandler {
 
-	protected void sendBinary(WebSocketSession session, String msg) {
-		if (StrUtil.isEmpty(msg)) {
-			return;
-		}
-		if (!session.isOpen()) {
-			// 会话关闭不能发送消息 @author jzy 21-08-04
-			log.warn("回话已经关闭啦，不能发送消息：{}", msg);
-			return;
-		}
-		synchronized (session.getId()) {
-			BinaryMessage byteBuffer = new BinaryMessage(msg.getBytes());
-			try {
-				session.sendMessage(byteBuffer);
-			} catch (IOException e) {
-				log.error("发送消息失败:" + msg, e);
-			}
-		}
-	}
+    protected void sendBinary(WebSocketSession session, String msg) {
+        if (msg == null) {
+            return;
+        }
+        BinaryMessage byteBuffer = new BinaryMessage(msg.getBytes());
+        try {
+            SocketSessionUtil.send(session, byteBuffer);
+        } catch (IOException e) {
+            log.error("发送消息失败:" + msg, e);
+        }
+    }
 }